refactor: NewProductInfo, EditProductInfo를 APIRequestDataType으로 이동
