+++ DotoriMarket/DotoriMarket/Network/APIRequest/RequestDataType/EditProductInfo.swift
@@ -1,27 +1,11 @@
 //
-//  ProductInfo.swift
-//  OpenMarket
+//  EditProductInfo.swift
+//  DotoriMarket
 //
-//  Created by lily on 2022/01/06.
+//  Created by 1 on 2022/07/05.
 //
 
 import Foundation
-
-struct NewProductInfo: Encodable {
-    
-    let name: String
-    let descriptions: String
-    let price: Double
-    let currency: CurrencyResponse
-    let discountedPrice: Double?
-    let stock: Int?
-    let secret: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case discountedPrice = "discounted_price"
-        case name, descriptions, price, currency, stock, secret
-    }
-}
 
 struct EditProductInfo: Encodable {
